Code clean up.
Removed the line sensor delay code. It was not helping.

Rearranged the stuff before setup() to put the tweakable stuff like fixed delays, motor speeds, analog thresholds, etc at the beginning, the constructor calls together, utility routines together, etc.

Cleaned up the blocking of the run modes to make them easier to find.

This compiles, and should be fine, but I did not test it yet.