A few tweaks...
Added code to straighten out the robot if it approaches a crossing line at an angle. This works unless the angle is too much. Crossing a double line at the outside of a turn is still problematic, but the course has been modified to eliminate these situations.

Modified the crossing line delays. Added 100 mS to the window for detecting the second line of a double line, making it 300 mS. Cut the blanking delay for detecting the next single line so we don't miss the crossing line on the other side of a road intersection.

Things pretty much work on my home test track.

Next steps...

* Counting lines or double lines to detect the end of the course.
* try using the analog outputs from the line sensors, maybe to do some rudimentary PID, or at least play with hysteresis.
* test on the real track.